Add classifiers documentation
