+++ docs/classifiers.rst
@@ -1,0 +1,22 @@
+.. _models:
+
+torchhd.classifiers
+=======================
+
+.. currentmodule:: torchhd.classifiers
+
+.. autosummary::
+    :toctree: generated/
+    :template: class.rst
+
+    Classifier
+    Vanilla
+    AdaptHD
+    OnlineHD
+    RefineHD
+    NeuralHD
+    DistHD
+    LeHDC
+    CompHD
+    SparseHD
+    QuantHD
