+++ torchhd/__init__.py
@@ -25,7 +25,7 @@ import torchhd.functional as functional
 import torchhd.embeddings as embeddings
 import torchhd.structures as structures
 import torchhd.models as models
-import torchhd.classify as classify
+import torchhd.classifiers as classifiers
 import torchhd.memory as memory
 import torchhd.datasets as datasets
 import torchhd.utils as utils
@@ -93,7 +93,7 @@ __all__ = [
     "embeddings",
     "structures",
     "models",
-    "classify",
+    "classifiers",
     "memory",
     "datasets",
     "utils",
