+++ torchhd/models.py
@@ -161,101 +161,6 @@ class Centroid(nn.Module):
         self.weight.index_add_(0, target, alpha1 * input, alpha=lr)
         self.weight.index_add_(0, pred, alpha2 * input, alpha=lr)
 
-    @torch.no_grad()
-    def add_adjust(
-        self,
-        input: Tensor,
-        target: Tensor,
-        count: int,
-        similarity_sum: float,
-        error_count: int,
-        error_similarity_sum: float,
-        lr: float = 1.0,
-    ) -> None:
-        logit = self(input)
-        predx = torch.topk(logit, 2)
-        pred = torch.tensor([predx.indices[0][0]])
-        is_wrong = target != pred
-
-        alpha = 1 - (abs(predx[0][0][0]) - abs(predx[0][0][1]))
-
-        similarity_sum += logit.max(1).values.item()
-        count += 1
-        if error_count == 0:
-            val = similarity_sum / count
-        else:
-            val = error_similarity_sum / error_count
-        if is_wrong.sum().item() == 0:
-            if logit.max(1).values.item() < val:
-                self.weight.index_add_(0, target, lr * alpha * input)
-            return
-
-        error_count += 1
-        error_similarity_sum += logit.max(1).values.item()
-
-        logit = logit[is_wrong]
-        input = input[is_wrong]
-        target = target[is_wrong]
-        pred = pred[is_wrong]
-        alpha1 = 1.0 - logit.gather(1, target.unsqueeze(1))
-        alpha2 = logit.gather(1, pred.unsqueeze(1)) - 1
-
-        self.weight.index_add_(0, target, lr * alpha1 * alpha * input)
-        self.weight.index_add_(0, pred, lr * alpha2 * alpha * input)
-
-    @torch.no_grad()
-    def add_sparse(
-        self, input: Tensor, target: Tensor, weight_sparse, lr: float = 1.0, iter=0
-    ) -> None:
-        if iter == 0:
-            logit = self(input)
-        else:
-            logit = functional.dot_similarity(input, self.weight_sparse)
-        pred = logit.argmax(1)
-        is_wrong = target != pred
-
-        if is_wrong.sum().item() == 0:
-            return
-
-        input = input[is_wrong]
-        target = target[is_wrong]
-        pred = pred[is_wrong]
-
-        if iter == 0:
-            self.weight.index_add_(0, target, lr * input)
-            self.weight.index_add_(0, pred, lr * -input)
-        else:
-            weight_sparse.index_add_(0, target, lr * input)
-            weight_sparse.index_add_(0, pred, lr * -input)
-
-    def quantized_similarity(self, input, model, weight_quant):
-        if model == "binary":
-            return functional.hamming_similarity(input, weight_quant).float()
-        elif model == "ternary":
-            return functional.dot_similarity(input, weight_quant)
-
-    def add_quantize(
-        self,
-        input: Tensor,
-        target: Tensor,
-        weight_quant,
-        lr: float = 1.0,
-        model="binary",
-    ) -> None:
-        logit = self.quantized_similarity(input, model, weight_quant)
-        pred = logit.argmax(1)
-        is_wrong = target != pred
-
-        if is_wrong.sum().item() == 0:
-            return
-
-        input = input[is_wrong]
-        target = target[is_wrong]
-        pred = pred[is_wrong]
-
-        self.weight.index_add_(0, target, lr * input)
-        self.weight.index_add_(0, pred, lr * -input)
-
     def normalize(self, eps=1e-12) -> None:
         """Transforms all the class prototype vectors into unit vectors.
